move mockChatsData to different file and import it from there
