+++ client/src/components/chats/mock_messages.js
@@ -1,0 +1,68 @@
+const chatsMockData = [
+  {
+    id: '3',
+    messages: [
+      {
+        content: 'Finished styling the ChatRoom page and testing it with dummy data from the messages store.',
+        sender_id: '61cdd39a5a14f24e4f2f89c7',
+        receiver_id: '3',
+        timestamp: 1652361968,
+      },
+    ],
+  },
+  {
+    id: '6',
+    messages: [
+      {
+        content: 'Look over chat array, sort, return latest message.',
+        sender_id: '',
+        receiver_id: '6',
+        timestamp: 1652362500,
+      },
+      {
+        content: 'Alright I will look over chat array, sort, and return latest message.',
+        sender_id: '61cdd39a5a14f24e4f2f89c7',
+        receiver_id: '6',
+        timestamp: 1652362533,
+      },
+
+    ],
+  },
+  {
+    id: '33',
+    messages: [
+      {
+        content: 'Define user state as variable te value of which is going to be hwatever the user.',
+        sender_id: '33',
+        receiver_id: '61cdd39a5a14f24e4f2f89c7',
+        timestamp: 1652261968,
+      },
+      {
+        content: 'You always say tomorrow and it is never done.',
+        sender_id: '33',
+        receiver_id: '61cdd39a5a14f24e4f2f89c73',
+        timestamp: 1652271970,
+      },
+      {
+        content: 'Ok thank you for all the details I should implement that tomorrow.',
+        sender_id: '61cdd39a5a14f24e4f2f89c7',
+        receiver_id: '33',
+        timestamp: 1652271971,
+      },
+
+    ],
+  },
+  {
+    id: '66',
+    messages: [
+      {
+        content: "Retrieve profile pic and name from user store using the frriends ID received from message router.'",
+        sender_id: '61cdd39a5a14f24e4f2f89c7',
+        receiver_id: '66',
+        timestamp: 1652185568,
+      },
+    ],
+  },
+]
+
+export { chatsMockData }
