+++ client/src/styles/FriendsStartAChat.module.css
@@ -16,6 +16,5 @@
 .friend__username{
     font-size: 18px;
     font-weight: 600;
-    width: 169px;
     margin: 0 10px;
 }
