+++ client/src/styles/LatestChat.module.css
@@ -1,7 +1,3 @@
-.latest-chats {
-    
-}
-
 .latest-chat {
     display: flex;
     align-items: center;
